build: compile build info into mysqld instead of sql_main
Build info changing on every build (because of timestamp) caused many
relinks. Move it higher in the tree, to reduce the number of relinks.

Before: villagesql_common -> sql_main -> libsql_main.a (75 links)
After: mysqld -> 59 links on macOS; 1 on Linux.

Saves 5-10s on no-change build on my mac.

1 wrinkle: sysvar registration. Now our static initialization is via
a file-scope object compiled directly into the executable (so linker
won't drop it.)

Also, tighten build-info test. It redacted every field so would have passed
against a server with no metadata at all.
